requirements.txt
agent-framework[azure]>=1.0.0rc3
agent-framework-orchestrations>=1.0.0b260304
azure-cosmos>=4.9.0
azure-identity>=1.19.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
